+++ src/main/java/qupath/ext/djl/ui/DjlEngineCommand.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package qupath.ext.djl;
+package qupath.ext.djl.ui;
 
 import java.nio.file.Files;
+import java.util.ResourceBundle;
 import java.util.Set;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
@@ -24,6 +25,11 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import ai.djl.util.cuda.CudaUtils;
+import javafx.geometry.Pos;
+import javafx.scene.Cursor;
+import javafx.scene.Node;
+import javafx.scene.text.TextAlignment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -46,12 +52,11 @@ import javafx.scene.control.ContentDisplay;
 import javafx.scene.control.Label;
 import javafx.scene.control.Separator;
 import javafx.scene.control.Tooltip;
-import javafx.scene.effect.DropShadow;
 import javafx.scene.layout.GridPane;
-import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
+import qupath.ext.djl.DjlTools;
 import qupath.fx.dialogs.Dialogs;
 import qupath.fx.utils.GridPaneUtils;
 import qupath.lib.common.GeneralTools;
@@ -71,6 +76,8 @@ public class DjlEngineCommand {
 	private static final String TITLE = "Deep Java Library engines";
 	
 	private QuPathGUI qupath;
+
+	private static final ResourceBundle bundle = ResourceBundle.getBundle("qupath.ext.djl.ui.strings");
 	
 	private enum EngineStatus {
 		UNAVAILABLE, UNKNOWN, PENDING, AVAILABLE, FAILED
@@ -88,12 +95,13 @@ public class DjlEngineCommand {
 			);
 	
 	/**
-	 * Several engines definitely don't work on Apple Silicon.
+	 * Several engines definitely don't work on Apple Silicon (for now anyway).
 	 */
 	private static Set<String> UNSUPPORTED_APPLE_SILICON = Set.of(
 			DjlTools.ENGINE_PADDLEPADDLE,
 			DjlTools.ENGINE_TENSORFLOW,
-			DjlTools.ENGINE_LIGHTGBM			
+			DjlTools.ENGINE_TFLITE,
+			DjlTools.ENGINE_LIGHTGBM
 			);
 	
 	/**
@@ -108,44 +116,48 @@ public class DjlEngineCommand {
 	
 	private void init() {
 		var pane = new GridPane();
+		pane.getStylesheets().add(DjlEngineCommand.class.getResource("styles.css").toExternalForm());
 		pane.setHgap(5);
 		pane.setVgap(5);
 		pane.setPadding(new Insets(10));
 		int row = 0;
-		
-		var label = new Label(
-				"Download Deep Java Library engines for inference."
-				);
-		label.setWrapText(true);
-		GridPaneUtils.addGridRow(pane, row++, 0, null, label, label);
-		var label2 = new Label(
-				"These will be stored in the local directories shown."
-				);
-		label2.setWrapText(true);
-		GridPaneUtils.addGridRow(pane, row++, 0, null, label2, label2);
 
-		var label3 = new Label(
-				"Please check the original engine's website for any licensing information.\n"
-				);
-		label3.setWrapText(true);
-		GridPaneUtils.addGridRow(pane, row++, 0, null, label3, label3);
+		var label = createCenteredLabelForString(
+				String.format(bundle.getString("overview.description"), Engine.getDjlVersion())
+		);
+		GridPaneUtils.addGridRow(pane, row++, 0, null, label, label, label);
+		var label2 = createCenteredLabelForKey("overview.directories");
+		GridPaneUtils.addGridRow(pane, row++, 0, null, label2, label2, label2);
+		var label3 = createCenteredLabelForKey("overview.licensing");
+		GridPaneUtils.addGridRow(pane, row++, 0, null, label3, label3, label3);
+
+		var separator = new Separator(Orientation.HORIZONTAL);
+		GridPaneUtils.addGridRow(pane, row++, 0, null, separator, separator, separator);
+
+		var labelCuda = createCenteredLabelForString(getCudaString());
+		GridPaneUtils.addGridRow(pane, row++, 0, null, labelCuda, labelCuda, labelCuda);
 
 		for (var name : Engine.getAllEngines()) {
-			var separator = new Separator(Orientation.HORIZONTAL);
-			GridPaneUtils.addGridRow(pane, row++, 0, null, separator, separator);
+			separator = new Separator(Orientation.HORIZONTAL);
+			GridPaneUtils.addGridRow(pane, row++, 0, null, separator, separator, separator);
 			GridPaneUtils.setToExpandGridPaneWidth(separator);
 			
 			var status = available.computeIfAbsent(name, n -> new SimpleObjectProperty<>(EngineStatus.UNKNOWN));
-			
+
 			var path = Utils.getEngineCacheDir(name);
-			
-			String tooltip = "Engine: " + name;
-			var labelName = new Label(name);
+
+			String tooltip = String.format(bundle.getString("tooltip.engine"), name);
+			String labelNameText = name;
+			if (name.equals(Engine.getDefaultEngineName())) {
+				labelNameText = String.format(bundle.getString("label.defaultName"), name);
+			}
+			var labelName = new Label(labelNameText);
 			labelName.setStyle("-fx-font-weight: bold;");
 			GridPaneUtils.addGridRow(pane, row++, 0, tooltip, labelName, labelName);
-			
+
 			var labelPath = new Label(path.toString());
-			var btnDownload = new Button("Download");
+			var btnDownload = new Button();
+			btnDownload.setText(bundle.getString("button.download"));
 			btnDownload.setPrefWidth(60.0);
 			long timeoutMillis = 500L;
 			btnDownload.setOnAction(e -> checkEngineStatus(name, status, timeoutMillis, false));
@@ -155,17 +167,16 @@ public class DjlEngineCommand {
 			btnDownload.textProperty().bind(Bindings.createStringBinding(() -> {
 				switch (status.get()) {
 				case AVAILABLE:
-					return "Available";
+					return bundle.getString("button.download.available");
 				case FAILED:
-					return "Try again";
+					return bundle.getString("button.download.failed");
 				case PENDING:
-					return "Pending";
-				case UNKNOWN:
-					return "Check / Download";
+					return bundle.getString("button.download.pending");
 				case UNAVAILABLE:
-					return "Download";
+					return bundle.getString("button.download.unavailable");
+				case UNKNOWN:
 				default:
-					return name + " download status is unknown";				
+					return bundle.getString("button.download.unknown");
 				}
 			}, status));
 			
@@ -188,45 +199,46 @@ public class DjlEngineCommand {
 			btnDownload.setTooltip(tooltipDownload);
 			
 			var circle = new Circle(6.0);
-			circle.setEffect(new DropShadow(2, 1, 1, Color.grayRgb(128, 0.5)));
-			circle.styleProperty().bind(Bindings.createStringBinding(() -> {
-				switch (status.get()) {
-				case AVAILABLE:
-					return "-fx-fill: lightgreen; -fx-opacity: 0.8";
-				case UNAVAILABLE:
-					return "-fx-fill: -fx-text-base-color; -fx-opacity: 0.5;";
-				case FAILED:
-					return "-fx-fill: darkred; -fx-opacity: 0.8";
-				case PENDING:
-					return "-fx-fill: yellow; -fx-opacity: 0.8";
-				case UNKNOWN:
-				default:
-					return "-fx-fill: orange; -fx-opacity: 0.8";
-				}
-			}, status));
-			var labelPathLabel = new Label("Path: ");
-			labelPathLabel.setGraphic(circle);
+			updateIconFromStatus(status.get(), circle);
+
+
+			var labelPathLabel = createLabelForKey("label.path");
 			labelPathLabel.setLabelFor(labelPath);
 			labelPathLabel.setContentDisplay(ContentDisplay.LEFT);
 
-			var tip = new Tooltip();
-			status.addListener((v, o, n) -> {
-				if (Files.isDirectory(path)) {
-					tip.setText("Double-click to open engine path");
-				} else {
-					labelPath.setStyle("-fx-opacity: 0.5;");
-					tip.setText("Engine path does not exist");	
-				}				
-			});
-			labelPathLabel.setTooltip(tip);
+			var tooltipPath = new Tooltip();
+			labelPath.setTooltip(tooltipPath);
+			labelPath.setCursor(Cursor.HAND);
 			labelPath.setOnMouseClicked(e -> {
 				if (e.getClickCount() == 2 && Files.isDirectory(path))
 					GuiTools.browseDirectory(path.toFile());
 				else
 					logger.debug("Cannot open {} - directory does not exist", path);
 			});
-			GridPaneUtils.addGridRow(pane, row++, 0, null, labelPathLabel, labelPath);
-			GridPaneUtils.addGridRow(pane, row++, 0, null, btnDownload, btnDownload);
+
+			var labelVersion = new Label("");
+			labelVersion.setMaxWidth(Double.MAX_VALUE);
+
+			var labelVersionLabel = createLabelForKey("label.version");
+			labelVersionLabel.setLabelFor(labelVersion);
+			labelVersionLabel.setContentDisplay(ContentDisplay.LEFT);
+
+			status.addListener((v, o, n) -> {
+				if (Files.isDirectory(path)) {
+					tooltipPath.setText(bundle.getString("tooltip.openPath"));
+				} else {
+					labelPath.setStyle("-fx-opacity: 0.5;");
+					tooltipPath.setText(bundle.getString("tooltip.pathMissing"));
+				}
+				updateIconFromStatus(n, circle);
+				updateVersionFromStatus(n, name, labelVersion);
+			});
+
+			pane.add(circle, 0, row, 1, 2);
+
+			GridPaneUtils.addGridRow(pane, row++, 1, null, labelPathLabel, labelPath);
+			GridPaneUtils.addGridRow(pane, row++, 1, null, labelVersionLabel, labelVersion);
+			GridPaneUtils.addGridRow(pane, row++, 1, null, btnDownload, btnDownload);
 			
 			GridPaneUtils.setToExpandGridPaneWidth(labelName, labelPath, btnDownload);
 
@@ -240,14 +252,84 @@ public class DjlEngineCommand {
 		stage.initModality(Modality.APPLICATION_MODAL);
 		stage.setScene(new Scene(pane));
 	}
-	
+
+	private static void updateIconFromStatus(EngineStatus status, Node node) {
+		if (status == null)
+			status = EngineStatus.UNKNOWN;
+		node.getStyleClass().setAll("djl-engine-status", status.name().toString().toLowerCase());
+	}
+
+	private static void updateVersionFromStatus(EngineStatus status, String engineName, Label labelVersion) {
+		if (status == EngineStatus.AVAILABLE) {
+			try {
+				var engine = Engine.getEngine(engineName);
+				labelVersion.setText(engine.getVersion());
+				labelVersion.setTooltip(new Tooltip(engine.toString()));
+				return;
+			} catch (Exception e) {
+				logger.error("Error updating engine version: {}", e.getMessage(), e);
+			}
+		}
+		labelVersion.setText("");
+		labelVersion.setTooltip(null);
+	}
+
+
+	private static Label createLabelForKey(String key) {
+		var label = new Label(bundle.getString(key));
+		label.setWrapText(true);
+		return label;
+	}
+
+
+	private static Label createCenteredLabelForKey(String key) {
+		return createCenteredLabelForString(bundle.getString(key));
+	}
+
+	private static Label createCenteredLabelForString(String s) {
+		var label = new Label(s);
+		label.setWrapText(true);
+		label.setAlignment(Pos.CENTER);
+		label.setTextAlignment(TextAlignment.CENTER);
+		label.setMaxHeight(Double.MAX_VALUE);
+		label.setMaxWidth(Double.MAX_VALUE);
+		return label;
+	}
+
+
 	private void showDialog() {
 		if (stage.isShowing())
 			stage.toFront();
 		else
 			stage.show();
 	}
-	
+
+	/**
+	 * Get a user-friendly string describing the CUDA version and GPUs.
+	 * @return
+	 */
+	private static String getCudaString() {
+		if (CudaUtils.hasCuda()) {
+			var sb = new StringBuilder()
+					.append(bundle.getString("label.cudaVersion"))
+					.append(CudaUtils.getCudaVersionString());
+			for (int i = 0; i < CudaUtils.getGpuCount(); i++) {
+				sb.append("\n")
+						.append("  GPU ")
+						.append(i)
+						.append(" ")
+						.append(bundle.getString("label.computeCapability"))
+						.append(CudaUtils.getComputeCapability(i));
+			}
+			return sb.toString();
+		} else if (GeneralTools.isMac() && "aarch64".equals(System.getProperty("os.arch"))) {
+			return bundle.getString("cuda.appleSilicon");
+		} else {
+			return bundle.getString("cuda.notFound");
+		}
+	}
+
+
 	private void checkEngineStatus(String name, ObjectProperty<EngineStatus> status, long timeoutMillis, boolean doQuietly) {
 		// Request the engine in a background thread, triggering download if necessary
 		var pool = Executors.newSingleThreadExecutor(ThreadTools.createThreadFactory("djl-engine-request", true));